# api/dist/swagger.yaml
swagger: '2.0'
info:
  description: >-
    REST API for the <a href="https://fdc.nal.usda.gov/download-datasets.html>various public nutrient data  </a> released on the USDA Food Data Central site.
  version: 1.0.0
  title: Nutrient Data API
  contact:
    name: Gary Moore
    email: gary@littlebunch.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8000'
basePath: /
...
schemes:
  - http
  - https
paths:
...
  /v1/nutrients/food/{id}:
    get:
      tags:
        - developers
      summary: returns nutrient data elements for a food item
      description:
        Returns either the list of nutrient data items  or a single nutrient data item for a food item identified
        by it's fdc id.
      operationId: nutrient
      produces:
        -application/json
      parameters:
        - name: n
          in: query
          description: >-
            return only the nutrient data element identifed by the nutrient number in n
          required: false
          type: integer
        - name: id
          in: path
          description: >-
            FDC ID of the food for which nutrient data is being requested
          required: true
      responses:
        '200':
          description: a single NutrientData element or a list of elements
          schema:
            $ref: '#/definitions/BrowseNutrientDataResult'
        '404':
          description: no results found
...
definitions:
...
  BrowseNutrientDataResult:
    type: object
    properties:
      count:
        type: integer
        format: int32
        description: Number of records found by a browse or search request
        example: 10
      start:
        type: integer
        description: >-
          Starting point (offset) into the list returned by a browse or search request
        format: int32
        example: 0
      max:
        type: integer
        description: The number of items in the list returned by a browse or search request
        format: int32
        example: 50
        minimum: 1
        maximum: 150
      items:
        type: array
        items:
          $ref: '#/definitions/nutrientData'
...
  nutrientData:
    required:
      - nutrientNumber
      - unit
    properties:
      fdcId:
        type: string
        example: 167514
      nutrientNumber:
        type: integer
        format: uint
        example: 301
      nutrientName:
        type: string
        example: Energy
      unit:
        type: string
        example: kcal
      derivation:
        type: array
        items: 
          $ref: '#/definitions/derivation'
      valuePer100UnitServing:
        type: number
        format: float
        example: 10.4
      datapoints:
        type: integer
        format: uint
  derivation:
    description: describes how a nutrient value was derived.
    required:
      - id
    properties:
      code:
        type: string
        example: "A"
      description:
        type: string
        exampe: "Analytical"
      id:
        type: integer
        format: uint
...